Removed unused classes (mat2, mat3). Reduced the number of calls to the copy constructor in the math library during main render loop by ~60-70%.

+++ OpenGL_Game/OpenGL_Game/display.cpp
@@ -68,7 +68,7 @@ GLvoid Window::startOpenGL()
 GLfloat Window::prepareFrame()
 {
 	// Per-frame time logic
-	GLfloat currentFrame = glfwGetTime();
+	GLfloat currentFrame = (GLfloat) glfwGetTime();
 	deltaTime = currentFrame - lastFrame;
 	lastFrame = currentFrame;
